Add files via upload
Just Import cryptography module and use Fernet Class.

Fernet is a class from the cryptography library in Python. It is used to perform authenticated encryption of data. Authenticated encryption ensures that the encrypted data cannot be tampered with or modified without the encryption key being detected.
Fernet is a symmetric encryption algorithm based on the Advanced Encryption Standard (AES) and uses a key and an Initialization Vector (IV) for encryption and decryption. The Fernet class provides a simple and secure way to encrypt and decrypt data using authenticated encryption.