MasterMine
MasterMine is a fun and engaging color-guessing game implemented in Python. In this game, players attempt to guess a secret code composed of a combination of colors within a limited number of tries. The game provides feedback on the number of correct colors in the correct positions and the number of correct colors in incorrect positions, helping players refine their guesses.

Features:

Randomly generated secret code using a selection of colors.
User-friendly input for guessing the code.
Feedback mechanism to guide players in their attempts.
Limited number of attempts to increase the challenge.
How to Play:

Clone the repository to your local machine.
Run the mastermine.py script using Python.
Enter your guesses for the secret code when prompted.
Try to guess the code within the given number of attempts!
Technologies Used:
-Python